tweaks in response to psychedelicious review of PR
